dalalstreet-ai (1)/dalalstreet-ai/backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-multipart==0.0.9

# ── Auth & Security ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==43.0.1

# ── Database ──────────────────────────────────────────────────
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# ── Redis / Cache ─────────────────────────────────────────────
redis[asyncio]==5.0.8
celery[redis]==5.4.0

# ── Rate Limiting ─────────────────────────────────────────────
slowapi==0.1.9

# ── HTTP Client ───────────────────────────────────────────────
httpx==0.27.2
aiohttp==3.10.5

# ── Data Science / ML ─────────────────────────────────────────
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
ta==0.11.0           # Technical Analysis library

# ── AI ────────────────────────────────────────────────────────
anthropic==0.34.2

# ── Broker APIs ───────────────────────────────────────────────
upstox-python-sdk==2.11.0
kiteconnect==5.0.1

# ── Validation ────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# ── Monitoring ────────────────────────────────────────────────
prometheus-fastapi-instrumentator==7.0.0

# ── Utilities ─────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.2
loguru==0.7.2
